+++ lib/generators/rails_email_preview/update_previews_generator.rb
@@ -25,14 +25,14 @@ module RailsEmailPreview
       def mailer_class_body(mailer_class_name)
 <<-RUBY
 class #{mailer_class_name}Preview
-#{mailer_methods(mailer_class_name) * "\n\n"}
+#{(mailer_methods(mailer_class_name) * "\n\n").chomp}
 end
 RUBY
       end
 
       def mailer_methods(mailer_class_name)
         mailer_class = mailer_class_name.constantize
-        mailer_class.instance_methods(false).map do |m|
+        ::RailsEmailPreview::Preview.mail_methods(mailer_class).map do |m|
 <<-RUBY
   def #{m}
     #{mailer_class_name}.#{m.to_s} #{mailer_class.instance_method(m).parameters.map(&:second) * ', '}
@@ -40,6 +40,7 @@ RUBY
 RUBY
         end
       end
+
     end
   end
 end