+++ lib/rails_email_preview.rb
@@ -16,8 +16,14 @@ module RailsEmailPreview
   mattr_accessor :parent_controller
   self.parent_controller = '::ApplicationController'
 
-  # auto-loading configured in initializer
+  # preview class names
   mattr_accessor :preview_classes
+
+  # UI locale
+  mattr_accessor :locale
+
+  # default email locale
+  mattr_accessor :default_email_locale
 
   # send email button
   mattr_accessor :enable_send_email
@@ -38,6 +44,11 @@ module RailsEmailPreview
   }
 
   class << self
+    def preview_classes=(classes)
+      @preview_classes = classes
+      RailsEmailPreview::Preview.load_all(classes)
+    end
+
     def layout=(layout)
       [::RailsEmailPreview::ApplicationController, ::RailsEmailPreview::EmailsController].each { |ctrl| ctrl.layout layout }
       if layout && layout !~ %r(^rails_email_preview/)
@@ -46,9 +57,9 @@ module RailsEmailPreview
       end
     end
 
-    def run_before_render(mail, preview_class_name, mailer_action)
+    def run_before_render(mail, preview)
       (defined?(@hooks) && @hooks[:before_render] || []).each do |block|
-        block.call(mail, preview_class_name, mailer_action)
+        block.call(mail, preview)
       end
     end