+++ spec/dummy/config/application.rb
@@ -9,8 +9,13 @@ require 'rails_email_preview'
 
 module Dummy
   class Application < Rails::Application
-    config.i18n.available_locales = [:es, :en]
+    config.i18n.available_locales = [:es, :en, :de]
     config.i18n.default_locale = :en
+
+    # Rails 3
+    if config.assets.respond_to?(:enabled=)
+      config.assets.enabled = true
+    end
   end
 end